# Loading Peaks faster with complete range and synchronization functionality.