### Loading the peaks faster with a complete range.

**Changes**

- This PR changes the functionality back to FURY where we avoid calculating voxels that don't contain any peaks.
- Ensuring the sizes will still correspond to the actual image.

**Goals**

- This PR makes much faster than the current Horizon peaks loading time.

**Speed Up**

- Measuring the speedup on a Linux machine with below given specifications
![Screenshot from 2024-04-10 11-51-18](https://github.com/dipy/dipy/assets/39947025/13d471c2-0349-443e-98c7-0f92ee79136e)

- Measuring was done using peaks generated from stanford_hardi data.

| Before | After |
|--------|--------|
| 63s | 5s |