@skoudoro 2024-04-05
Hi @maharshi-gor,

Do you have benchmarks? to see how much faster