@maharshi-gor 2024-04-10
Hello @skoudoro, I have updated the description to show the system info and the performance boost.

I opened up peaks using the master branch and current branch to compare.